+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.3] - 2021-05-21
 ### Added
 - Simple calculation method for mosfet gate resistor (Paper PCIM 2006: D.Kübrich 'Invetsigation of Turn-Off Behaviour under the Assumption of Linear Capacitances')
 ### Bugfix
@@ -58,8 +60,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   the database
 - Matlab-Exporter
 
-[Unreleased]: https://github.com/upb-lea/Transistor_Database/compare/0.2.2...HEAD
-[0.2.1]: https://github.com/upb-lea/Transistor_Database/compare/0.2.1...0.2.2
+[Unreleased]: https://github.com/upb-lea/Transistor_Database/compare/0.2.3...HEAD
+[0.2.3]: https://github.com/upb-lea/Transistor_Database/compare/0.2.2...0.2.3
+[0.2.2]: https://github.com/upb-lea/Transistor_Database/compare/0.2.1...0.2.2
 [0.2.1]: https://github.com/upb-lea/Transistor_Database/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/upb-lea/Transistor_Database/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/upb-lea/Transistor_Database/compare/0.1.0...0.1.0